Log if we need to skip a CRTB/PRTB due to the user not existing
This feels like the safer option versus applying permissions that
none of the users we've collected actually have, even with the
GUID/DN matching. This situation should be relatively uncommon,
as Rancher usually cleans these up when users are deleted, but
with the GUID duplicate bug I'm not sure how successful that will
have been in practice. Best to be safe (and noisy)