For CRTB/PRTBs, rework error handling to gracefully retry
In particular, this treats internal errors (usually related to
webhook timeouts) as transient, and retries them with a little bit
of exponential backoff.

Furthermore, after reviewing some scenarios with Michael, we've
decided to consider non-internal errors from the webhook as
non-fatal in terms of continuing to process the individual user.
There are a few situations where old bindings to disabled templates
would otherwise block users from migrating, and this permits those
to have a better chance of overall success.